refactor(audio_input): remove the unreachable PipeWire input backends
`selected_pipewire_bridge_backend` is the only producer of
`PipewireBridgeBackendKind`, and `InputClockMode` has exactly three
variants, so the selection could only ever return `PwClientNode`
(upstream) or `PwStream` (dac/pipewire). The `PwExportedNode`,
`PwAdapter` and `PwFilter` arms were unreachable, and everything
reachable only through them was dead: `pipewire_exported.rs` entirely,
and all of `pipewire_legacy.rs` except two symbols.

Keeping them was not free. `live_input.rs` held a full copy of the
pw_stream capture path and its DRIVER pacing helpers, and that copy had
already drifted behind fixes made in the live `audio_input/pipewire.rs`:
it never re-read IEC958 pods on format change, and it still derived
transport frames from `pw_time` instead of the observed chunk size.
Silent rot in code nothing can execute.

Relocate the two symbols that were still live rather than keep a module
for them: `PipewireBridgeBackendKind` (reduced to its two reachable
variants, which now makes the selection exhaustive to the compiler)
moves to `pipewire.rs` next to the backend it selects, and
`clone_spa_pod_bytes` moves to `pipewire_pods.rs` next to the other pod
helpers, where `pipewire_client_node.rs` already imports from.

The investigation report that documents why these backends were
abandoned is kept, with a status note so a reader does not go looking
for code that no longer exists.

No behavioural change: both shipping backends are untouched.
