+++ omniphony-renderer/audio_input/src/lib.rs
@@ -6,10 +6,6 @@ pub mod pipewire;
 #[cfg(target_os = "linux")]
 pub mod pipewire_client_node;
 #[cfg(target_os = "linux")]
-pub mod pipewire_exported;
-#[cfg(target_os = "linux")]
-pub mod pipewire_legacy;
-#[cfg(target_os = "linux")]
 pub mod pipewire_pods;
 
 pub use control::{
